# Add /Qspectre compiler flags