Adding compiler options to guard against security vulnerabilities  (/Qspectre and /guard:cf). Part of SDL compliance activity.