@pranavsharma 2019-03-21
Should this option be enabled by default? Is there any reason we wouldn't want to enable this? Trying to see if we can eliminate this additional build option.