@jignparm 2019-03-21
> Should this option be enabled by default? Is there any reason we wouldn't want to enable this? Trying to see if we can eliminate this additional build option.

Only consideration is that users building on Windows will require Visual Studio with  KB 4338871 installed to enable the /Qspectre option (see https://aka.ms/Ofhn4c). This is not installed by default on community editions of VS or Azure VMs. So compiling OnnxRuntime out of the box would require most users to install the toolset.

We can set as the default option however, if we're looking to reduce number of build configurations.