@RyanUnderhill 2019-03-21
Is it known how Spectre could affect inference? Since we don't have any timing functions I'm not sure how someone could read protected memory by running a model.