@jignparm 2019-03-22
Regarding performance, the quote from the /Qspectre page mentions there should be negligible impact. It seems that loading a model could potentially be considered crossing a trust boundary, since it falls under "parses untrusted input or files" category. 

>**"The performance impact of /Qspectre has been seen to be negligible in several very large code bases, but there are no guarantees that performance of your code under /Qspectre remains unaffected"**