@pranavsharma 2019-03-22
Since this flag is going to be part of the officially released binaries, let's verify it doesn't affect the performance of most important 1P models.