@jignparm 2019-03-28
A brief summary of performance numbers for ResNet and TinyYolo.  The code base is exactly the same, one DLL is built using **--enable_qspectre**, another without it.

**/Qspectre** is not seem to be negatively impacting performance. It seems to result in slightly more efficient scoring code (at least for these two models).

with /Qspectre flag
-------------------

D:\onnxruntime\build\Windows\RelWithDebInfo\RelWithDebInfo>onnxruntime_perf_test.exe -s -r 1000 d:\models\test_resnet50\model.onnx perf.resnet.withspectre.txt
Setting thread pool size to 6
Total time cost:35.8138
Total iterations:1000
Average time cost:35.8138 ms


D:\onnxruntime\build\Windows\RelWithDebInfo\RelWithDebInfo>onnxruntime_perf_test.exe -s -r 1000 d:\models\test_tiny_yolov2\model.onnx perf.yolo.withspectre.txt
Setting thread pool size to 6
Total time cost:30.2484
Total iterations:1000
Average time cost:30.2484 ms


Without /Qspectre flag
----------------------

D:\onnxruntime\build\Windows\RelWithDebInfo\RelWithDebInfo>onnxruntime_perf_test.exe -s -r 1000 d:\models\test_resnet50\model.onnx perf.resnet.withoutspectre.txt
Setting thread pool size to 6
Total time cost:39.4769
Total iterations:1000
Average time cost:39.4769 ms

D:\onnxruntime\build\Windows\RelWithDebInfo\RelWithDebInfo>onnxruntime_perf_test.exe -s -r 1000 d:\models\test_tiny_yolov2\model.onnx perf.yolo.withoutspectre.txt
Setting thread pool size to 6
Total time cost:35.7952
Total iterations:1000
Average time cost:35.7952 ms