@pranavsharma 2019-03-28
Thanks Jignesh for the performance numbers. Given the numbers, should be enable it by default? If it's the default there's one less thing to worry about in build configs.