@jignparm 2019-03-28
> Thanks Jignesh for the performance numbers. Given the numbers, should be enable it by default? If it's the default there's one less thing to worry about in build configs.

Only consideration is that users building on Windows will require Visual Studio with KB 4338871 installed to enable the /Qspectre option (see https://aka.ms/Ofhn4c). This is not installed by default on community editions of VS or Azure VMs. So compiling OnnxRuntime out of the box would require most users to install the toolset.

If these are made mandatory, the installation steps Build.md can be updated (small change), but the primary issue is requiring users to install the spectre libraries even though they are not required. Preference is to keep it optional (for ease of compilation for users). If there are any strong thoughts otherwise, we can make these flags mandatory.