@pranavsharma 2019-03-28
Also, can you find out how much additional binary size does this add?