@jignparm 2019-03-28
The binary dll size is 4,855,808 bytes without spectre, and 4,882,944 bytes with spectre. So about 0.5% diff (27KB).