@jignparm 2019-03-29
If /Qspectre is specified without /Zi, BinSkim cannot analyze the binary file. Therefore need to set /Zi whenever /Qspectre is set. /Zi is usually set by default in Debug build and RelWithDebInfo, but not for Release build.