@snnn 2019-03-29
The why don't use the Release build?